+++ src/main/java/api/m2/file/clients/identity/IdentityClient.java
@@ -10,6 +10,7 @@ import api.m2.file.clients.identity.response.UserLookupResponse;
 import api.m2.file.clients.identity.response.UserMe;
 import api.m2.file.clients.identity.response.WorkspaceAdded;
 import api.m2.file.clients.identity.response.WorkspaceInvitationDTO;
+import api.m2.file.clients.identity.response.WorkspaceSentInvitationDTO;
 import api.m2.file.clients.identity.response.WorkspaceMemberDTO;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -68,5 +69,11 @@ public interface IdentityClient {
 
     @PatchExchange("/v1/invitations")
     void acceptRejectInvitation(@RequestBody AcceptRejectInvitationDTO body);
+
+    @GetExchange("/v1/invitations/sent")
+    List<WorkspaceSentInvitationDTO> getSentInvitations();
+
+    @DeleteExchange("/v1/invitations/{invitationId}")
+    void cancelInvitation(@PathVariable Long invitationId);
 
 }