+++ src/main/java/api/m2/file/enums/InvitationStatus.java
@@ -3,5 +3,6 @@ package api.m2.file.enums;
 public enum InvitationStatus {
     PENDING,
     ACCEPTED,
-    REJECTED
+    REJECTED,
+    CANCELLED
 }